Remove now unexistant include.

+++ Playground/main.cpp
@@ -6,7 +6,6 @@
 #include "../JSONParser.h"
 #include "../FlatMap.h"
 #include "../FlatSet.h"
-#include "../HashTable.h"
 using namespace ARLib;
 
 constexpr static size_t n_of_strings = 400;
@@ -28,11 +27,10 @@ void erase_all(FlatSet<String>& set) {
         Printer::print("len {} was {}", s.size(), set.remove(s) ? "removed" : "not removed");
     }
 }
-
 void fill_map(FlatMap<String, int>& set) {
     for (size_t i = 1; i < n_of_strings; ++i) {
         String s{ i, 'a' };
-        set.insert(move(s), i);
+        set.insert(move(s), static_cast<int>(i));
     }
 }
 void search_map(const FlatMap<String, int>& set) {
@@ -47,7 +45,6 @@ void erase_map(FlatMap<String, int>& set) {
         Printer::print("len {} was {}", s.size(), set.remove(s) ? "removed" : "not removed");
     }
 }
-
 int main() {
     FlatSet<String> set{};
     fill_set(set);
@@ -56,12 +53,10 @@ int main() {
     erase_all(set);
     for (const auto& str : set) { Printer::print("{}", str); }
 
-
     FlatMap<StringView, int> map2{};
     map2.insert("hello"_sv, 1);
     map2.find("hello");
-    
-    
+
     FlatMap<String, int> map{};
     fill_map(map);
     search_map(map);
